Refactor Files and Folders, Introduce Controllers
Performed file and folder refactoring to improve code organization and maintainability. Created a new 'controllers' folder to centralize the logic for handling product and cart routes.
